// src/app/models/teacherProfile.model.ts
import { TeacherQualificationModel } from "./teacherQualification.model";

export interface TeacherProfileModel {
  TeacherProfileId: number;
  UserId: number;
  CreatedOn: string;
  ModifiedOn: string;
  PersonalizedUrl: string;
  VideoIntroUrl: string;
  Introduction: string;
  Repertoir: string;
  City: string;
  ZipCode: string;
  TimeZone: string;
  DisplayName: string;
  StudioName: string;
  ShortIntroduction: string;
  AvailabilityAcceptingNew: boolean;
  AvailabilityAutoAccept: boolean;
  AvailabilityOfferSolo: boolean;
  AvailabilityOfferGroup: boolean;
  AcceptingAgeLow?: number;
  AcceptingAgeHigh?: number;
  VocalStylePop: boolean;
  VocalStyleClassical: boolean;
  VocalStyleMusicalTheater: boolean;
  VocalStyleGospel: boolean;
  VocalStyleHipHop: boolean;
  VocalStyleOpera: boolean;
  VocalStyleJazz: boolean;
  VocalStyleFolk: boolean;
  VocalStyleRock: boolean;
  VocalStyleWorld: boolean;
  VocalType: string;
  MusicalWorshipBaptist: boolean;
  MusicalWorshipCatholic: boolean;
  MusicalWorshipHindi: boolean;
  MusicalWorshipLutheran: boolean;
  MusicalWorshipNonDenom: boolean;
  MusicalWorshipPresbyterian: boolean;
  MusicalWorshipBuddhist: boolean;
  MusicalWorshipEpiscopal: boolean;
  MusicalWorshipJewish: boolean;
  MusicalWorshipMethodist: boolean;
  MusicalWorshipMuslim: boolean;
  MusicalWorshipUnitarian: boolean;
  MusicalWorshipOther: string;
  SocialMediaFacebook: string;
  SocialMediaTwitter: string;
  SocialMediaInstagram: string;
  SocialMediaLinkedIn: string;
  SocialMediaYouTube: string;
  SocialMediaPinterest: string;
  WebsiteUrl?: string;
  User?: null;
  TeacherQualifications?: Array<TeacherQualificationModel> | null;
}